Move the entire Forge buildSrc into this
Warning: this is **super duper incredibly scuffed!** But I did it in a way that Gradle likes. Dependencies are totally messed up, too, but it works. This is due for a huge cleanup, ***but it works.***

That being said, I may need to do some extra testing with the Forge repo without the submodules initialized. The idea is that the Forge repo should work with the ForgeDevPlugin repo uninitialized (this is by checking if `forgedev/settings.gradle` exists). But the ForgeDev 7 side of the move is done with this commit.