Temporarily disable Shadow plugin
Forge buildscript uses some exposed API of the buildSrc classes I just moved (mostly ASM). Workaround for now is to just disable the shadow jar so the workspace can continue to work. Again, that will be part of a later cleanup. ForgeDev does not need to be as polished as ForgeGradle (which will resume development soon).